# openapi/harness-core-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto-agent HTTP API",
    "version": "v1",
    "description": "Versioned SaaS HTTP/SSE contract for the routes registered by pkg/server. Tenant and subject ownership are resolved from the authenticated Principal; cross-tenant session access is intentionally reported as 404. This pre-GA v1 contract accepts additive fields and routes, while semantic or breaking changes require a new reviewed contract revision. Operations marked x-harness-contract-first are approved control-plane contracts whose server registration is verified before they can be considered implemented. Optional adapters return the documented 501 error when they are not configured."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/v1/admin/profiles/{id}": {"get": {"operationId": "getAdminProfile", "tags": ["admin", "profiles"], "summary": "Resolve one profile's effective view at an owned scope", "x-harness-contract-first": true, "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ProfileID"}, {"$ref": "#/components/parameters/ScopeQuery"}], "responses": {"200": {"description": "The effective profile snapshot for the requested scope. A tenant operator may only request its tenant scope or a descendant; an account administrator may request any valid scope.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminProfileResponse"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}}}, "put": {"operationId": "putAdminProfile", "tags": ["admin", "profiles"], "summary": "Create or replace one scoped AgentProfileLayer", "x-harness-contract-first": true, "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ProfileID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminProfileWriteRequest"}}}}, "responses": {"200": {"description": "The recomposed effective profile after the scoped layer is accepted.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminProfileResponse"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}}}},
...
    "/v1/admin/rag/projection": {"get": {"operationId": "ragProjectionStats", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {"type": "http", "scheme": "bearer", "description": "Built-in account token, or a deployment-injected Authenticator. The resolved Principal supplies tenant, subject, scope, grants, and role; handlers enforce ownership and administrator boundaries. Development header fallback is disabled in production."},
      "RunnerBearer": {"type": "http", "scheme": "bearer", "description": "Private runner token; it cannot access administrative runner task routes."}
    },
    "parameters": {
...
      "ProfileID": {"name": "id", "in": "path", "required": true, "description": "A safe 1-128-character profile ID. Built-in bare IDs such as general and namespaced IDs are both valid.", "schema": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"}},
...
      "ScopeQuery": {"name": "scope", "in": "query", "required": true, "description": "ScopePath as its JSON array encoding, URL encoded; e.g. [{\"kind\":\"tenant\",\"id\":\"acme\"}].", "schema": {"type": "string", "minLength": 1}},
...
    },
    "responses": {
      "Success": {"description": "Implemented successful response; exact payload varies by operation."},
      "Error": {"description": "Uniform JSON error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "string"}}, "additionalProperties": false},
...
      "ScopeRef": {"type": "object", "required": ["kind", "id"], "properties": {"kind": {"type": "string", "enum": ["global", "deployment", "product", "tenant", "workspace", "user", "session", "run"]}, "id": {"type": "string", "minLength": 1}}, "additionalProperties": false},
      "ScopePath": {"type": "array", "minItems": 1, "items": {"$ref": "#/components/schemas/ScopeRef"}, "description": "Root-to-leaf ownership path. This is the real JSON encoding of core.ScopePath, not its slash-delimited diagnostic String form."},
      "ModelSelection": {"type": "object", "required": ["provider", "model"], "properties": {"provider": {"type": "string"}, "model": {"type": "string"}}, "additionalProperties": false},
      "PromptFragment": {"type": "object", "required": ["id", "section", "content"], "properties": {"id": {"type": "string"}, "section": {"type": "string", "enum": ["identity", "traits", "values", "goals", "style", "boundaries", "instructions"]}, "content": {"type": "string", "minLength": 1}, "priority": {"type": "integer"}}, "additionalProperties": false},
      "AgentProfileLayer": {"type": "object", "required": ["profile_id"], "description": "One scoped profile contribution. Its scope is carried by the surrounding admin request because core.AgentProfileLayer intentionally omits scope from its direct JSON form.", "properties": {"profile_id": {"type": "string", "minLength": 1}, "extends": {"type": "string"}, "name": {"type": "string"}, "description": {"type": "string"}, "model": {"$ref": "#/components/schemas/ModelSelection"}, "max_steps": {"type": "integer", "minimum": 1}, "max_tool_calls": {"type": "integer", "minimum": 1}, "add_capabilities": {"type": "array", "items": {"type": "string"}}, "remove_capabilities": {"type": "array", "items": {"type": "string"}}, "put_fragments": {"type": "array", "items": {"$ref": "#/components/schemas/PromptFragment"}}, "remove_fragments": {"type": "array", "items": {"type": "string"}}, "metadata": {"type": "object", "additionalProperties": {"type": "string"}}}, "additionalProperties": false},
      "ResolvedPromptFragment": {"allOf": [{"$ref": "#/components/schemas/PromptFragment"}, {"type": "object", "required": ["source"], "properties": {"source": {"$ref": "#/components/schemas/ScopePath"}}}]},
      "AgentProfileSnapshot": {"type": "object", "required": ["id", "profile_id", "created_at", "scope", "name", "model", "max_steps", "capabilities", "fragments"], "properties": {"id": {"type": "string"}, "profile_id": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "scope": {"$ref": "#/components/schemas/ScopePath"}, "name": {"type": "string"}, "description": {"type": "string"}, "model": {"$ref": "#/components/schemas/ModelSelection"}, "max_steps": {"type": "integer"}, "max_tool_calls": {"type": "integer"}, "capabilities": {"type": "array", "items": {"type": "string"}}, "fragments": {"type": "array", "items": {"$ref": "#/components/schemas/ResolvedPromptFragment"}}, "metadata": {"type": "object", "additionalProperties": {"type": "string"}}}},
      "AdminProfileWriteRequest": {"type": "object", "required": ["scope", "layer"], "properties": {"scope": {"$ref": "#/components/schemas/ScopePath"}, "layer": {"$ref": "#/components/schemas/AgentProfileLayer"}}, "additionalProperties": true},
      "AdminProfileResponse": {"type": "object", "required": ["scope", "layer", "effective"], "properties": {"scope": {"$ref": "#/components/schemas/ScopePath"}, "layer": {"$ref": "#/components/schemas/AgentProfileLayer"}, "effective": {"$ref": "#/components/schemas/AgentProfileSnapshot"}}},
...
    }
  }
}